Moved views to view_model.
